+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## [Unreleased]
 
+- Add support for uint8, uint16, uint32 and uint64
+
 ## [0.1.3] - 2025-09-13
 
-- Add support for all int8, int16, int32, float16 and float32
+- Add support for int8, int16, int32, float16 and float32
 
 ## [0.1.2] - 2025-09-12